+++ packages/veui/demo/cases/Checkbox.vue
@@ -2,10 +2,16 @@
   <article>
     <h1><code>&lt;veui-checkbox&gt;</code></h1>
     <p>
-      <veui-checkbox value="normal" v-model="picked1">正常状态</veui-checkbox>
+      <veui-checkbox>未绑定数据</veui-checkbox>
     </p>
     <p>
-      <veui-checkbox value="small" ui="small" v-model="picked2">正常状态</veui-checkbox>
+      <veui-checkbox v-model="current" true-value="已选" false-value="未选">当前数据：</veui-checkbox> {{ current }}
+    </p>
+    <p>
+      <veui-checkbox v-model="picked1">正常状态</veui-checkbox>
+    </p>
+    <p>
+      <veui-checkbox ui="small" v-model="picked2">正常状态</veui-checkbox>
     </p>
     <p>
       <veui-checkbox v-model="picked3">选中状态</veui-checkbox>
@@ -50,6 +56,7 @@ export default {
   },
   data () {
     return {
+      current: '已选',
       picked1: false,
       picked2: false,
       picked3: true,