+++ packages/veui/src/components/Radio.vue
@@ -3,7 +3,7 @@
     'veui-radio': true,
     'veui-disabled': realReadonly || realDisabled
   }" :ui="ui">
-  <input type="radio" v-bind="attrs" @change="$emit('change', $event.target.checked)">
+  <input type="radio" v-bind="attrs" @change="localChecked = $event.target.checked">
   <span class="veui-radio-box"></span>
   <span class="veui-radio-label"><slot></slot></span>
 </label>
@@ -11,26 +11,44 @@
 
 <script>
 import { input } from '../mixins'
-import { pick } from 'lodash'
 
 export default {
   name: 'veui-radio',
   mixins: [input],
-  props: {
-    ui: String,
-    value: null,
-    checked: Boolean
-  },
   model: {
     prop: 'checked',
     event: 'change'
   },
+  props: {
+    ui: String,
+    value: {
+      type: null,
+      default: true
+    },
+    checked: null
+  },
+  data () {
+    return {
+      localChecked: this.checked
+    }
+  },
   computed: {
     attrs () {
-      let attrs = pick(this.$props, 'checked')
-      attrs.name = this.realName
-      attrs.disabled = this.realDisabled || this.realReadonly
-      return attrs
+      return {
+        checked: this.localChecked,
+        name: this.realName,
+        disabled: this.realDisabled || this.realReadonly
+      }
+    }
+  },
+  watch: {
+    checked (value) {
+      this.localChecked = value
+    },
+    localChecked (value) {
+      if (this.checked !== value) {
+        this.$emit('change', value)
+      }
     }
   }
 }