+++ packages/veui/src/components/Tooltip.vue
@@ -44,6 +44,10 @@ export default {
     open: {
       type: Boolean,
       default: false
+    },
+    interactive: {
+      type: Boolean,
+      default: true
     }
   },
   data () {
@@ -82,8 +86,12 @@ export default {
         handler: this.closeHandler,
         refs: this.targetNode,
         trigger: this.localTrigger.close,
-        delay: this.hideDelay
+        delay: this.hideDelay,
+        excludeSelf: !this.interactive
       }
+    },
+    realOpen () {
+      return this.localOpen && !!this.targetNode
     }
   },
   watch: {
@@ -96,6 +104,9 @@ export default {
       if (this.open !== val) {
         this.$emit('update:open', val)
       }
+    },
+    target () {
+      this.localOpen = true
     }
   },
   methods: {
@@ -128,9 +139,12 @@ export default {
     return (
       <veui-overlay
         target={this.targetNode}
-        open={this.localOpen}
+        open={this.realOpen}
         options={this.overlay}
-        overlayClass={this.mergeOverlayClass('veui-tooltip-box')}>
+        overlayClass={this.mergeOverlayClass({
+          'veui-tooltip-box': true,
+          'veui-tooltip-box-transparent': !this.interactive
+        })}>
         <div class="veui-tooltip" ui={this.ui} {...{directives}}>
           <div class="veui-tooltip-content">
             { this.$slots.default }